# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EstateOS TAGRe — Property Management Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard summary data
  - name: properties
    description: Property management
  - name: tenants
    description: Tenant management
  - name: maintenance
    description: Maintenance tickets
  - name: leases
    description: Lease management
  - name: landlords
    description: Landlord management and churn scoring
  - name: campaigns
    description: Vendor email campaigns
  - name: data-sources
    description: External data source integration status
  - name: ai-config
    description: AI configuration and provider management
  - name: audit
    description: Audit log of AI actions

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard KPI summary
      responses:
        "200":
          description: Dashboard summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /dashboard/charts:
    get:
      operationId: getDashboardCharts
      tags: [dashboard]
      summary: Get chart data for dashboard
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardCharts"

  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"

  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyDetail"

  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"

  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetail"

  /maintenance:
    get:
      operationId: listMaintenanceTickets
      tags: [maintenance]
      summary: List maintenance tickets
      parameters:
        - name: priority
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of maintenance tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceTicket"

  /maintenance/{id}:
    get:
      operationId: getMaintenanceTicket
      tags: [maintenance]
      summary: Get maintenance ticket details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicketDetail"

  /maintenance/{id}/approve:
    post:
      operationId: approveMaintenanceAction
      tags: [maintenance]
      summary: Approve AI-suggested maintenance action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveActionBody"
      responses:
        "200":
          description: Action approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  /leases:
    get:
      operationId: listLeases
      tags: [leases]
      summary: List all leases
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lease"

  /landlords:
    get:
      operationId: listLandlords
      tags: [landlords]
      summary: List landlords with churn scores
      responses:
        "200":
          description: List of landlords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Landlord"

  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List vendor email campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"

  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign with AI draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign with draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"

  /campaigns/{id}/approve:
    post:
      operationId: approveCampaignDraft
      tags: [campaigns]
      summary: Approve AI-drafted vendor email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveDraftBody"
      responses:
        "200":
          description: Draft approved and queued for send
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  /data-sources:
    get:
      operationId: listDataSources
      tags: [data-sources]
      summary: List all external data source integrations and their status
      responses:
        "200":
          description: Data source statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataSource"

  /data-sources/{id}/ingest:
    post:
      operationId: triggerIngestion
      tags: [data-sources]
      summary: Manually trigger data ingestion for a source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ingestion triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"

  /data-sources/events:
    get:
      operationId: listIngestionEvents
      tags: [data-sources]
      summary: List recent ingestion events across all sources
      responses:
        "200":
          description: Recent events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IngestionEvent"

  /ai-config/providers:
    get:
      operationId: listAiProviders
      tags: [ai-config]
      summary: List AI provider configurations
      responses:
        "200":
          description: Providers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiProvider"

  /ai-config/tasks:
    get:
      operationId: listAiTaskAssignments
      tags: [ai-config]
      summary: List AI task to provider assignments
      responses:
        "200":
          description: Task assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiTaskAssignment"

  /ai-config/spend:
    get:
      operationId: getAiSpend
      tags: [ai-config]
      summary: Get AI spend and telemetry summary
      responses:
        "200":
          description: Spend data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSpendSummary"

  /audit:
    get:
      operationId: listAuditLog
      tags: [audit]
      summary: List AI audit log entries
      parameters:
        - name: task
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    DashboardSummary:
      type: object
      properties:
        totalProperties:
          type: integer
        activeLeases:
          type: integer
        openMaintenance:
          type: integer
        pendingApprovals:
          type: integer
        rentRollValue:
          type: number
        arrearsCount:
          type: integer
        vacancyRate:
          type: number
        aiDraftsToday:
          type: integer
        landlordChurnRisk:
          type: integer
        monthlyRevenue:
          type: number
      required: [totalProperties, activeLeases, openMaintenance, pendingApprovals, rentRollValue, arrearsCount, vacancyRate, aiDraftsToday, landlordChurnRisk, monthlyRevenue]

    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        timestamp:
          type: string
        status:
          type: string
        aiGenerated:
          type: boolean
        urgent:
          type: boolean
      required: [id, type, title, description, timestamp, status, aiGenerated]

    DashboardCharts:
      type: object
      properties:
        rentCollectionByMonth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              collected:
                type: number
              outstanding:
                type: number
            required: [month, collected, outstanding]
        maintenanceByCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required: [category, count]
        propertyStatusBreakdown:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required: [status, count]
        landlordChurnTrend:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              atRisk:
                type: integer
              saved:
                type: integer
            required: [month, atRisk, saved]
      required: [rentCollectionByMonth, maintenanceByCategory, propertyStatusBreakdown, landlordChurnTrend]

    Property:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        suburb:
          type: string
        type:
          type: string
        status:
          type: string
        weeklyRent:
          type: number
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        tenantName:
          type: string
        landlordName:
          type: string
        leaseExpiry:
          type: string
        maintenanceOpen:
          type: integer
        imageUrl:
          type: string
      required: [id, address, suburb, type, status, weeklyRent, bedrooms, bathrooms, maintenanceOpen]

    PropertyDetail:
      allOf:
        - $ref: "#/components/schemas/Property"
        - type: object
          properties:
            leaseHistory:
              type: array
              items:
                type: object
            maintenanceHistory:
              type: array
              items:
                type: object
            landlordNotes:
              type: string

    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyAddress:
          type: string
        leaseStart:
          type: string
        leaseEnd:
          type: string
        weeklyRent:
          type: number
        arrears:
          type: number
        status:
          type: string
        avatarInitials:
          type: string
      required: [id, name, email, phone, propertyAddress, leaseStart, leaseEnd, weeklyRent, arrears, status]

    TenantDetail:
      allOf:
        - $ref: "#/components/schemas/Tenant"
        - type: object
          properties:
            paymentHistory:
              type: array
              items:
                type: object
            maintenanceRequests:
              type: array
              items:
                type: object

    MaintenanceTicket:
      type: object
      properties:
        id:
          type: string
        propertyAddress:
          type: string
        tenantName:
          type: string
        category:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        submittedAt:
          type: string
        aiTriaged:
          type: boolean
        estimatedCost:
          type: number
        suggestedAction:
          type: string
        pendingApproval:
          type: boolean
      required: [id, propertyAddress, tenantName, category, description, priority, status, submittedAt, aiTriaged, pendingApproval]

    MaintenanceTicketDetail:
      allOf:
        - $ref: "#/components/schemas/MaintenanceTicket"
        - type: object
          properties:
            aiAnalysis:
              type: string
            confidenceScore:
              type: number
            suggestedVendors:
              type: array
              items:
                type: object
            ownerPreferences:
              type: string
            draftOwnerSms:
              type: string
            draftTenantAck:
              type: string

    ApproveActionBody:
      type: object
      properties:
        action:
          type: string
        notes:
          type: string
      required: [action]

    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
      required: [success, message, timestamp]

    Lease:
      type: object
      properties:
        id:
          type: string
        propertyAddress:
          type: string
        tenantName:
          type: string
        landlordName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        weeklyRent:
          type: number
        status:
          type: string
        daysUntilExpiry:
          type: integer
        documentParsed:
          type: boolean
        complianceStatus:
          type: string
      required: [id, propertyAddress, tenantName, landlordName, startDate, endDate, weeklyRent, status, daysUntilExpiry, documentParsed, complianceStatus]

    Landlord:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        propertiesCount:
          type: integer
        portfolioValue:
          type: number
        churnScore:
          type: integer
        churnRisk:
          type: string
        topRiskFactors:
          type: array
          items:
            type: string
        recommendedIntervention:
          type: string
        lastContact:
          type: string
        avatarInitials:
          type: string
      required: [id, name, email, phone, propertiesCount, portfolioValue, churnScore, churnRisk, topRiskFactors, recommendedIntervention, lastContact]

    Campaign:
      type: object
      properties:
        id:
          type: string
        propertyAddress:
          type: string
        vendorName:
          type: string
        agentName:
          type: string
        listingDate:
          type: string
        priceGuide:
          type: string
        enquiries:
          type: integer
        inspections:
          type: integer
        offers:
          type: integer
        status:
          type: string
        draftStatus:
          type: string
        weekNumber:
          type: integer
      required: [id, propertyAddress, vendorName, agentName, listingDate, priceGuide, enquiries, inspections, offers, status, draftStatus, weekNumber]

    CampaignDetail:
      allOf:
        - $ref: "#/components/schemas/Campaign"
        - type: object
          properties:
            emailSubject:
              type: string
            emailBody:
              type: string
            confidenceScore:
              type: number
            aiReasoning:
              type: string
            priceRecommendation:
              type: string
            comparables:
              type: array
              items:
                type: object

    ApproveDraftBody:
      type: object
      properties:
        editedSubject:
          type: string
        editedBody:
          type: string
        notes:
          type: string
      required: []

    DataSource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        system:
          type: string
        integrationMode:
          type: string
        status:
          type: string
        lastSync:
          type: string
        nextSync:
          type: string
        recordsIngested:
          type: integer
        eventsToday:
          type: integer
        errorCount:
          type: integer
        isLive:
          type: boolean
        phase:
          type: string
        logoColor:
          type: string
      required: [id, name, description, system, integrationMode, status, lastSync, recordsIngested, eventsToday, errorCount, isLive, phase]

    IngestionJob:
      type: object
      properties:
        jobId:
          type: string
        sourceId:
          type: string
        status:
          type: string
        startedAt:
          type: string
        estimatedRecords:
          type: integer
      required: [jobId, sourceId, status, startedAt, estimatedRecords]

    IngestionEvent:
      type: object
      properties:
        id:
          type: string
        sourceId:
          type: string
        sourceName:
          type: string
        eventType:
          type: string
        payload:
          type: string
        timestamp:
          type: string
        status:
          type: string
        recordsAffected:
          type: integer
      required: [id, sourceId, sourceName, eventType, payload, timestamp, status, recordsAffected]

    AiProvider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        vendor:
          type: string
        status:
          type: string
        region:
          type: string
        models:
          type: array
          items:
            type: string
        monthlyCost:
          type: number
        callCount:
          type: integer
        avgLatencyMs:
          type: integer
        successRate:
          type: number
        isDefault:
          type: boolean
      required: [id, name, vendor, status, region, models, monthlyCost, callCount, avgLatencyMs, successRate, isDefault]

    AiTaskAssignment:
      type: object
      properties:
        id:
          type: string
        taskName:
          type: string
        taskLabel:
          type: string
        primaryProvider:
          type: string
        primaryModel:
          type: string
        fallbackProvider:
          type: string
        fallbackModel:
          type: string
        costPerCall:
          type: number
        callsToday:
          type: integer
        avgQualityScore:
          type: number
        containsPii:
          type: boolean
      required: [id, taskName, taskLabel, primaryProvider, primaryModel, fallbackProvider, fallbackModel, costPerCall, callsToday, avgQualityScore, containsPii]

    AiSpendSummary:
      type: object
      properties:
        totalThisMonth:
          type: number
          description: AUD
        budget:
          type: number
        spendByTask:
          type: array
          items:
            type: object
            properties:
              task:
                type: string
              spend:
                type: number
              calls:
                type: integer
            required: [task, spend, calls]
        spendByProvider:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              spend:
                type: number
            required: [provider, spend]
        dailySpend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              amount:
                type: number
            required: [date, amount]
      required: [totalThisMonth, budget, spendByTask, spendByProvider, dailySpend]

    AuditEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
        userId:
          type: string
        userName:
          type: string
        task:
          type: string
        taskLabel:
          type: string
        provider:
          type: string
        model:
          type: string
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        costAud:
          type: number
        latencyMs:
          type: integer
        status:
          type: string
        workflowId:
          type: string
        action:
          type: string
      required: [id, timestamp, userId, userName, task, taskLabel, provider, model, inputTokens, outputTokens, costAud, latencyMs, status, workflowId, action]