# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EstateOS TAGRe — Property Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  type: string;
  title: string;
  description: string;
  timestamp: string;
  status: string;
  aiGenerated: boolean;
  urgent?: boolean;
}