# lib/api-zod/src/generated/types/auditEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EstateOS TAGRe — Property Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEntry {
  id: string;
  timestamp: string;
  userId: string;
  userName: string;
  task: string;
  taskLabel: string;
  provider: string;
  model: string;
  inputTokens: number;
  outputTokens: number;
  costAud: number;
  latencyMs: number;
  status: string;
  workflowId: string;
  action: string;
}