# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EstateOS TAGRe — Property Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProperties: number;
  activeLeases: number;
  openMaintenance: number;
  pendingApprovals: number;
  rentRollValue: number;
  arrearsCount: number;
  vacancyRate: number;
  aiDraftsToday: number;
  landlordChurnRisk: number;
  monthlyRevenue: number;
}