# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EstateOS TAGRe — Property Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: string;
  address: string;
  suburb: string;
  type: string;
  status: string;
  weeklyRent: number;
  bedrooms: number;
  bathrooms: number;
  tenantName?: string;
  landlordName?: string;
  leaseExpiry?: string;
  maintenanceOpen: number;
  imageUrl?: string;
}