# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EstateOS TAGRe — Property Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: string;
  name: string;
  email: string;
  phone: string;
  propertyAddress: string;
  leaseStart: string;
  leaseEnd: string;
  weeklyRent: number;
  arrears: number;
  status: string;
  avatarInitials?: string;
}