# lib/api-zod/src/generated/types/tenantDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EstateOS TAGRe — Property Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Tenant } from "./tenant";
import type { TenantDetailMaintenanceRequestsItem } from "./tenantDetailMaintenanceRequestsItem";
import type { TenantDetailPaymentHistoryItem } from "./tenantDetailPaymentHistoryItem";

export type TenantDetail = Tenant & {
  paymentHistory?: TenantDetailPaymentHistoryItem[];
  maintenanceRequests?: TenantDetailMaintenanceRequestsItem[];
};